Fixed tk bug in the show results tab in the GUI. This happened on some systems where X11 threading was not working correctly. Also adds some features to mdt_math_img.

+++ mdt/cli_scripts/mdt_math_img.py
@@ -25,8 +25,8 @@ class MathImg(BasicShellApplication):
         description = textwrap.dedent("""
             Evaluate an expression on a set of images.
 
-            This is meant to quickly convert/combine one or two maps with a mathematical expression. The expression
-            can be any valid python string.
+            This is meant to quickly convert/combine one or two maps with a mathematical expression.
+            The expression can be any valid python expression.
 
             The input list of images are loaded as numpy arrays and stored in the array 'input', or equivalent, 'i'.
             Next, the expression is evaluated using the input images and the result is stored in the indicated file.
@@ -39,17 +39,28 @@ class MathImg(BasicShellApplication):
 
             You can use every single alphabetic character except for the i since that one is reserved for the array.
 
-            The module numpy is also available using the variable 'np'. This allows expressions like:
+            The module numpy is available under 'np' and some functions of MDT under 'mdt'.
+            This allows expressions like:
                 np.mean(np.concatenate(i, axis=3), axis=3)
             to get the mean value per voxel of all the input images.
+
+            You could also do something like:
+                list(map(lambda f: np.mean(mdt.create_roi(f, i[-1])), i[0:-1]))
+            to get for every map loaded the mean value over the given mask (assuming that the last map loaded is
+            the mask).
+
+            If no output file is specified and the output is of dimension 2 or lower we print the output directly
+            to the console.
         """)
         description += self._get_citation_message()
 
         epilog = textwrap.dedent("""
             Examples of use:
-                mdt-math-img fiso.nii ficvf.nii '(1-input[0]) * i[1]' Wic.w.nii.gz
-                mdt-math-img fiso.nii ficvf.nii '(1-a) * b' Wic.w.nii.gz
-                mdt-math-img *.nii.gz 'np.mean(np.concatenate(i, axis=3), axis=3)' output.nii.gz
+                mdt-math-img fiso.nii ficvf.nii '(1-input[0]) * i[1]' -o Wic.w.nii.gz
+                mdt-math-img fiso.nii ficvf.nii '(1-a) * b' -o Wic.w.nii.gz
+                mdt-math-img *.nii.gz 'np.mean(np.concatenate(i, axis=3), axis=3)' -o output.nii.gz
+                mdt-math-img FA.nii.gz 'np.mean(a)'
+                mdt-math-img images*.nii.gz mask.nii 'list(map(lambda f: np.mean(mdt.create_roi(f, i[-1])), i[0:-1]))'
         """)
 
         parser = argparse.ArgumentParser(description=description, epilog=epilog,
@@ -61,9 +72,9 @@ class MathImg(BasicShellApplication):
         parser.add_argument('expr', metavar='expr', type=str,
                             help="The expression to evaluate.")
 
-        parser.add_argument('output_file',
+        parser.add_argument('-o', '--output_file',
                             action=mdt.shell_utils.get_argparse_extension_checker(['.nii', '.nii.gz', '.hdr', '.img']),
-                            help='the base output name').completer = \
+                            help='the output file, if not set nothing is written').completer = \
             FilesCompleter(['nii', 'gz', 'hdr', 'img'], directories=False)
 
         parser.add_argument('--verbose', '-v', action='store_true', help="Verbose, prints runtime information")
@@ -71,10 +82,11 @@ class MathImg(BasicShellApplication):
         return parser
 
     def run(self, args):
-        output_file = os.path.realpath(args.output_file)
+        if args.output_file is not None:
+            output_file = os.path.realpath(args.output_file)
 
-        if os.path.isfile(output_file):
-            os.remove(output_file)
+            if os.path.isfile(output_file):
+                os.remove(output_file)
 
         file_names = []
         for file in args.input_files:
@@ -84,7 +96,7 @@ class MathImg(BasicShellApplication):
             print('')
 
         images = [mdt.load_volume(dwi_image)[0] for dwi_image in file_names]
-        context_dict = {'input': images, 'i': images}
+        context_dict = {'input': images, 'i': images, 'np': np, 'mdt': mdt}
         alpha_chars = list('abcdefghjklmnopqrstuvwxyz')
 
         for ind, image in enumerate(images):
@@ -99,9 +111,9 @@ class MathImg(BasicShellApplication):
             print('')
             print("Evaluating: '{expr}'".format(expr=args.expr))
 
-        output = eval(args.expr, {'np': np}, context_dict)
+        output = eval(args.expr, context_dict)
 
-        if len(output.shape) == 3:
+        if isinstance(output, np.ndarray) and len(output.shape) == 3:
             if args.verbose:
                 print('')
                 print('Output is 3d, reshaping to 4d')
@@ -110,12 +122,19 @@ class MathImg(BasicShellApplication):
         else:
             if args.verbose:
                 print('')
-                print('Output shape: {shape}'.format(shape=str(output.shape)))
+                if isinstance(output, np.ndarray):
+                    print('Output shape: {shape}'.format(shape=str(output.shape)))
+                else:
+                    print('Output is not matrix like')
+                print('Output: ')
+                print('')
+            print(output)
 
         if args.verbose:
             print('')
 
-        mdt.write_image(output_file, output, mdt.load_nifti(file_names[0]).get_header())
+        if args.output_file is not None:
+            mdt.write_image(output_file, output, mdt.load_nifti(file_names[0]).get_header())
 
 
 if __name__ == '__main__':
